fix: moves ResourceTypes to storage module
