+++ server/src/main/java/com/epam/aidial/core/server/data/ListPublishedResourcesRequest.java
@@ -1,5 +1,6 @@
 package com.epam.aidial.core.server.data;
 
+import com.epam.aidial.core.storage.resource.ResourceTypes;
 import lombok.Data;
 
 import java.util.Set;
