+++ server/src/main/java/com/epam/aidial/core/server/service/InvitationService.java
@@ -4,7 +4,6 @@ import com.epam.aidial.core.config.ResourceAccessType;
 import com.epam.aidial.core.server.data.Invitation;
 import com.epam.aidial.core.server.data.InvitationCollection;
 import com.epam.aidial.core.server.data.InvitationsMap;
-import com.epam.aidial.core.server.data.ResourceTypes;
 import com.epam.aidial.core.server.data.SharedResource;
 import com.epam.aidial.core.server.security.ApiKeyGenerator;
 import com.epam.aidial.core.server.security.EncryptionService;
@@ -12,6 +11,8 @@ import com.epam.aidial.core.server.util.ProxyUtil;
 import com.epam.aidial.core.server.util.ResourceDescriptorFactory;
 import com.epam.aidial.core.storage.exception.ResourceNotFoundException;
 import com.epam.aidial.core.storage.resource.ResourceDescriptor;
+import com.epam.aidial.core.storage.resource.ResourceType;
+import com.epam.aidial.core.storage.resource.ResourceTypes;
 import com.epam.aidial.core.storage.service.ResourceService;
 import io.vertx.core.json.JsonObject;
 import lombok.extern.slf4j.Slf4j;
@@ -264,7 +265,7 @@ public class InvitationService {
         }
         String location = parts[0] + ResourceDescriptor.PATH_SEPARATOR + parts[1] + ResourceDescriptor.PATH_SEPARATOR;
         String bucket = encryptionService.encrypt(location);
-        ResourceTypes resourceType = ResourceTypes.of(parts[2]);
+        ResourceType resourceType = ResourceTypes.of(parts[2]);
         return ResourceDescriptorFactory.fromDecoded(resourceType, bucket, location, INVITATION_RESOURCE_FILENAME);
     }
